style: Класс ErrorResponse перенесен в свой собственный java класс в пакете model
