+++ src/main/java/ru/yandex/practicum/filmorate/model/ErrorResponse.java
@@ -1,0 +1,20 @@
+package ru.yandex.practicum.filmorate.model;
+
+public class ErrorResponse {
+
+    private String error;
+    private String description;
+
+    public ErrorResponse(String error, String description) {
+        this.error = error;
+        this.description = description;
+    }
+
+    public String getError() {
+        return error;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+}
